client replay: move all client replay-related code into addon
