--- test/mitmproxy/proxy/protocol/test_http_replay.py
@@ -1,1 +1,0 @@
-# TODO: write tests
